libceph: encapsulate blocks of connection code
There's a funny block of code in try_write() that sets up for a
connection request to a server and puts a ceph connection in a
connected state if it is not already set up.

There's a similar block of code in try_read() that reads in the
response from a server to such a connection request.

Encapsulate both of those into functions that do a better job of
identifying their purpose.

This patch involves little more than code motion, to simplify
review.

Signed-off-by: Alex Elder <elder@inktank.com>