libceph: make use of distinct CONNECTING state
Small but significant change that exploits and enforces the fact
that NEGOTIATING is now distinct from CONNECTING.  Done as a
separate commit for verification purposes.

The CONNECTING bit is also now cleared in con_close_socket(), to
reflect that we have ended the connection phase once the socket
has been closed.  (A new one will begin when needed.)

Signed-off-by: Alex Elder <elder@inktank.com>