libceph: distinguish two phases of connect sequence
Currently a ceph connection enters a "CONNECTING" state when it
begins the process of (re-)connecting with its peer.  The first
step of this is for both parties to exchange a banner, then to
exchange their IP addresses (client sends its own to server, and
server sends its own and client's back to client).

Once that is complete, the client sends the server a connect
message along with an optional block of authorization data.
During this phase, an additional NEGOTIATING bit is set in
the ceph connection's state--though the CONNECTING bit/state
remains.

Finally, if the server provides a positive response to the client's
connect message, the connection is "alive" and usable.  At this
point both NEGOTIATING and CONNECTING are dropped from the ceph
connection state, and it enteres CONNECTED state.

Rather than have the CONNECTING state continue while the NEGOTIATING
bit is set, interpret these two phases as distinct states.  In other
words, when NEGOTIATING is set, clear CONNECTING.  That way only
one of them will be active at a time.

Signed-off-by: Alex Elder <elder@inktank.com>