libceph: define and use DISCONNECTED state
Rather than just using the SOCK_CLOSED flag, define a new state
DISCONNECTED to represent the condition of having received notice
that a socket has closed (or waiting for close).

We might get multiple close events reported on the socket, so allow
the CLOSING connection socket state to transition to itself.

Signed-off-by: Alex Elder <elder@inktank.com>