libceph: treat TAG_CLOSE as DISCONNECTED
When a TAG_CLOSE is received, it is an indication that the other end
of a ceph connection is about to close the connection.  Right now
this is recorded by transitioning the connection directly to CLOSED
state, but a better solution is to transition to DISCONNECTED state.
That state is now entered when a socket notification indicates the
socket is shutting down.  We can expand the interpretation of
DISCONNECTED to mean the connection was shut down externally (either
at the socket level or the ceph connection level).

Change the handling in con_work() in the DISCONNECTED state so it
occurs earlier--first thing--so we don't unnecessarily back off
a socket that's basically already closed (or should be).

Signed-off-by: Alex Elder <elder@inktank.com>