libceph: close connection on protocol failure
If a problem occurs during the connection message exchange portion
of the connect sequence, fail_protocol() gets called.  That resets
the connection and sets the state of the connection to CLOSED, but
doesn't really do a complete close of the connection.

Call ceph_con_close() from fail_protocol() instead.  That function
achieves the same affect on the connection state, resets the
connection, and closes the socket as well.

Signed-off-by: Alex Elder <elder@inktank.com>