libceph: don't close socket in OPENING state
The only way a socket enters OPENING state is via ceph_con_open().
The only times ceph_con_open() are called are:
  - In fs/ceph/mds_client.c:register_session(), where it occurs
     right after a call to ceph_con_init().
  - In fs/ceph/mds_client.c:send_mds_reconnect().  This is
    called in two places.
    - In fs/ceph/mds_client.c:check_new_map(), it is called
      after a call to ceph_con_close() (either directly or via
      fs/ceph/mds_client.c:__unregister_session()).
    - Or in fs/ceph/mds_client.c:peer_reset(), which is also
      only called after a call to ceph_con_close().
  - In net/ceph/mon_client.c:__open_session(), where it's called
    right after a call to ceph_con_init().
  - In net/ceph/osd_client.c:__reset_osd(), right after a call
    to ceph_con_close().
  - In net/ceph/osd_client.c:__map_request(), right after a call
    to create_osd(), which includes a call to ceph_con_init().

After a call to ceph_con_init(), the state of a ceph connection is
CLOSED, and its socket pointer is null.  Similarly, after a call to
ceph_con_close(), the state of the connection is CLOSED, the
underlying socket is closed, and the connection's socket pointer is
null.

Therefore, there is no reason to call con_close_socket() when a
connection is found to be in OPENING state in con_work().

Signed-off-by: Alex Elder <elder@inktank.com>