libceph: disconnects require close of socket
For connections that allow lossy transmission, errors due to lost
data are tolerated and ignored.  The code that implements this is
embedded in ceph_fault().  However ceph_fault() is also used to
handle faults due to notification that the underlying socket is
no longer connected, and in this case we need to close the socket
just like a "normal" connection error.

Change ceph_fault() so LOSSYTX channels only ignore errors that
are due to data loss, not socket connection loss.

Signed-off-by: Alex Elder <elder@inktank.com>